Tweaks to formatting, esRADT final

+++ src/interfaces/radiant-interfaces/IPoolHelper.sol
@@ -10,32 +10,19 @@ interface IPoolHelper {
 
     function zapWETH(uint256 amount) external returns (uint256);
 
-    function zapTokens(
-        uint256 _wethAmt,
-        uint256 _rdntAmt
-    ) external returns (uint256);
+    function zapTokens(uint256 _wethAmt, uint256 _rdntAmt) external returns (uint256);
 
-    function quoteFromToken(
-        uint256 tokenAmount
-    ) external view returns (uint256 optimalWETHAmount);
+    function quoteFromToken(uint256 tokenAmount) external view returns (uint256 optimalWETHAmount);
 
-    function getLpPrice(
-        uint rdntPriceInEth
-    ) external view returns (uint256 priceInEth);
+    function getLpPrice(uint256 rdntPriceInEth) external view returns (uint256 priceInEth);
 
-    function getReserves()
-        external
-        view
-        returns (uint256 rdnt, uint256 weth, uint256 lpTokenSupply);
+    function getReserves() external view returns (uint256 rdnt, uint256 weth, uint256 lpTokenSupply);
 
     function getPrice() external view returns (uint256 priceInEth);
 }
 
 interface IBalancerPoolHelper is IPoolHelper {
-    function initializePool(
-        string calldata _tokenName,
-        string calldata _tokenSymbol
-    ) external;
+    function initializePool(string calldata _tokenName, string calldata _tokenSymbol) external;
 }
 
 interface IUniswapPoolHelper is IPoolHelper {
