+++ src/interfaces/radiant-interfaces/IWETH.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.15;
 
 interface IWETH {
-    function balanceOf(address) external returns (uint);
+    function balanceOf(address) external returns (uint256);
 
     function deposit() external payable;
 
@@ -10,16 +10,9 @@ interface IWETH {
 
     function approve(address guy, uint256 wad) external returns (bool);
 
-    function transferFrom(
-        address src,
-        address dst,
-        uint256 wad
-    ) external returns (bool);
+    function transferFrom(address src, address dst, uint256 wad) external returns (bool);
 
-    function transfer(address to, uint value) external returns (bool);
+    function transfer(address to, uint256 value) external returns (bool);
 
-    function allowance(
-        address owner,
-        address spender
-    ) external returns (uint256);
+    function allowance(address owner, address spender) external returns (uint256);
 }
