+++ src/policies/VaultRewards.sol
@@ -16,43 +16,28 @@ contract Rewards is Policy, RolesConsumer {
     uint256 public epochRewards;
     uint256 public constant SCALAR = 1e6;
 
-    function configureDependencies()
-        external
-        override
-        returns (Keycode[] memory dependencies)
-    {
+    function configureDependencies() external override returns (Keycode[] memory dependencies) {
         dependencies = new Keycode[](1);
         dependencies[0] = toKeycode("ROLES");
         ROLES = ROLESv1(getModuleAddress(dependencies[0]));
     }
 
-    function requestPermissions()
-        external
-        pure
-        override
-        returns (Permissions[] memory requests)
-    {
+    function requestPermissions() external pure override returns (Permissions[] memory requests) {
         requests = new Permissions[](1);
     }
 
-    function setLeveragers(
-        ERC4626[] memory _leveragers
-    ) external onlyRole("ADMIN") {
+    function setLeveragers(ERC4626[] memory _leveragers) external onlyRole("ADMIN") {
         for (uint256 i = 0; i < _leveragers.length; i++) {
             leveragers.push(ERC4626(_leveragers[i]));
         }
     }
 
-    function setRewardWeights(
-        uint256[] memory _weights
-    ) external onlyRole("ADMIN") {
+    function setRewardWeights(uint256[] memory _weights) external onlyRole("ADMIN") {
         epochRewards = esRADT.balanceOf(address(this));
         // Log the number of shares
-        for (uint i = 0; i < leveragers.length; i++) {
+        for (uint256 i = 0; i < leveragers.length; i++) {
             uint256 poolRewards = (epochRewards * _weights[i]) / SCALAR;
-            rewardsPerShare[i] =
-                (poolRewards * SCALAR) /
-                leveragers[i].totalSupply();
+            rewardsPerShare[i] = (poolRewards * SCALAR) / leveragers[i].totalSupply();
         }
     }
 
@@ -69,10 +54,7 @@ contract Rewards is Policy, RolesConsumer {
         esRADT.transfer(msg.sender, rewards);
     }
 
-    function recoverERC20(
-        ERC20 _token,
-        uint256 _amount
-    ) external onlyRole("ADMIN") {
+    function recoverERC20(ERC20 _token, uint256 _amount) external onlyRole("ADMIN") {
         _token.transfer(msg.sender, _amount);
     }
 }
